# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  backend_route-fix-backup-20250603_161334_handlers.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
    type: object
...
  request.DashboardShareRequest:
    properties:
      permission_level:
        description: e.g., "view", "edit", "manage"
        enum:
        - view
        - edit
        - manage
        type: string
      shared_with_role_id:
        description: Pointer to allow null
        type: integer
      shared_with_user_id:
        description: Pointer to allow null
        type: integer
    required:
    - permission_level
    type: object
...
  response.DashboardListEntry:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_user_id:
        type: integer
      updated_at:
        type: string
    type: object
  response.DashboardListResponse:
    properties:
      dashboards:
        items:
          $ref: '#/definitions/response.DashboardListEntry'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total_count:
        type: integer
    type: object
...
  response.DashboardShareResponse:
    properties:
      created_at:
        type: string
      dashboard_id:
        type: string
      id:
        type: integer
      permission_level:
        type: string
      shared_with_role_id:
        type: integer
      shared_with_role_name:
        description: Populated if SharedWithRoleID is not nil
        type: string
      shared_with_user_id:
        type: integer
      shared_with_username:
        description: Populated if SharedWithUserID is not nil
        type: string
    type: object
...
  response.WidgetDataResponse:
    properties:
      data:
        description: Actual data structure varies by widget
      updated_at:
        type: string
      widget_id:
        type: string
    type: object
...
host: reksolindo.opuschamber.com
info:
  contact: {}
  description: Multi-tenant enterprise asset management system
  title: Reksolindo Enterprise Asset Management API
  version: "1.0"
paths:
...
  /dashboards:
    get:
      description: Retrieves a paginated list of dashboards owned by or shared with
        the current user.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of dashboards
          schema:
            $ref: '#/definitions/response.DashboardListResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List dashboards accessible to the user
      tags:
      - dashboards
  /dashboards/{dashboardId}/share:
    post:
      consumes:
      - application/json
      description: Shares a dashboard with a specified user or role.
      parameters:
      - description: Dashboard ID to share
        in: path
        name: id
        required: true
        type: string
      - description: Details of the share (user/role and permission level)
        in: body
        name: shareDetails
        required: true
        schema:
          $ref: '#/definitions/request.DashboardShareRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard shared successfully
          schema:
            type: object
        "400":
          description: Invalid Dashboard ID or request payload
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "403":
          description: Forbidden - Not authorized to share this dashboard
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "404":
          description: Dashboard, User, or Role not found
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Share a dashboard
      tags:
      - dashboards
  /dashboards/{dashboardId}/share/{shareId}:
    delete:
      consumes:
      - application/json
      description: Removes a specific share for a dashboard using the share ID.
      parameters:
      - description: Dashboard ID from which to unshare
        in: path
        name: id
        required: true
        type: string
      - description: The ID of the share record to delete
        in: path
        name: shareId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard unshared successfully
          schema:
            type: object
        "400":
          description: Invalid Dashboard ID or Share ID
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "403":
          description: Forbidden - Not authorized to unshare this dashboard
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "404":
          description: Dashboard or Share not found
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unshare a dashboard
      tags:
      - dashboards
  /dashboards/{dashboardId}/shares:
    get:
      description: Retrieves all share records for a specific dashboard. Only owner
        can view.
      parameters:
      - description: Dashboard ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of dashboard shares
          schema:
            items:
              $ref: '#/definitions/response.DashboardShareResponse'
            type: array
        "400":
          description: Invalid Dashboard ID
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "403":
          description: Forbidden - Not authorized to view shares
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "404":
          description: Dashboard not found
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get shares for a dashboard
      tags:
      - dashboards
  /dashboards/{dashboardId}/widgets/{widgetId}/data:
    get:
      consumes:
      - application/json
      description: Retrieves data for a widget on a specific dashboard.
      parameters:
      - description: Dashboard ID
        in: path
        name: id
        required: true
        type: string
      - description: Widget ID
        in: path
        name: widgetId
        required: true
        type: string
      - description: Additional widget parameters
        in: query
        name: params
        type: object
      produces:
      - application/json
      responses:
        "200":
          description: Widget data
          schema:
            $ref: '#/definitions/response.WidgetDataResponse'
        "400":
          description: Invalid Dashboard ID, Widget ID, or parameters
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "403":
          description: Forbidden - Access to dashboard/widget denied
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "404":
          description: Dashboard or Widget not found
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get data for a specific widget
      tags:
      - dashboards
...
schemes:
- https
- http
swagger: "2.0"